+++ examples/actix3_example/.env
@@ -1,0 +1,3 @@
+HOST=127.0.0.1
+PORT=8000
+DATABASE_URL="mysql://root:root@localhost/actix_example"
