+++ examples/actix3_example/entity/Cargo.toml
@@ -17,7 +17,7 @@ version = "^0.7.0"
 features = [
   "macros",
   "debug-print",
-  "runtime-actix-native-tls",
+  "runtime-async-std-native-tls",
   "sqlx-mysql",
   # "sqlx-postgres",
   # "sqlx-sqlite",
