+++ examples/actix3_example/static/css/style.css
