+++ examples/actix3_example/static/images/favicon.png
