+++ examples/actix3_example/templates/new.html.tera
