+++ examples/actix_example/README.md
@@ -1,6 +1,6 @@
 ![screenshot](Screenshot.png)
 
-# Actix with SeaORM example app
+# Actix 4 with SeaORM example app
 
 1. Modify the `DATABASE_URL` var in `.env` to point to your chosen database
